Add files via upload
ASSIGNMENT 01 QUESTIONS
Q 01. Create and configure a topology of distance vector routing information protocol RIPv2,
taking A is the numbers of routers and using any classless networks. (A must be the
sum of last two digits of your EP Number for e.g: if your Ep Number is “ EP1849023 ”
then the Value of A = 2 + 3 = 5.
Therefore, A = 5 for your problem solving).
Q 02. To make broadcast domain on layer 2 using Virtual Lan, create A numbers of VLANs.
(A must be the sum of last two digits of your EP Number for e.g: if your Ep Number is
“EP1849023 ” then the Value of A = 2 + 3 = 5.
Therefore, A = 5 for your problem solving).